rock-psoc-data-pipeline/requirements.txt
pandas==2.0.3
requests==2.31.0
python-dotenv==1.0.0
supabase==1.0.4
openpyxl==3.1.2